:recycle: (916): extract extract_participants into _shared and measure it in the participant-naming evaluation
